:bug:Fix: fix backtest option query error 전략 페이지에서 쿼리가 전달되지 않는 문제 해결
